Trying to improve tariff forms
